requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.3
python-dotenv==0.19.2
Flask-Mail==0.9.1
qrcode==7.4.2
Pillow==10.0.0
gunicorn==20.1.0
